@@ -24,14 +24,12 @@
   "homepage": "https://github.com/andrewbranch/typedoc-loader#readme",
   "dependencies": {
     "loader-utils": "^1.1.0",
-    "mock-fs": "^4.4.2",
-    "typedoc": "^0.12.0"
+    "typedoc": "^0.13.0"
   },
   "devDependencies": {
-    "jest": "^22.0.1",
-    "json-loader": "^0.5.7",
+    "jest": "^23.6.0",
     "memory-fs": "^0.4.1",
-    "webpack": "^4.17.1"
+    "webpack": "^4.23.1"
   },
   "peerDependencies": {
     "typescript": "^3.0.0"
