@@ -1,8 +1,16 @@
+const { Application } = require('typedoc');
 const { getOptions } = require('loader-utils');
-const StringProgrammaticApplication = require('./string-programmatic-application');
 
 module.exports = function() {
-  const options = getOptions(this);
-  const app = new StringProgrammaticApplication(options);
-  return JSON.stringify(app.toJson(this.resourcePath));
+  const app = new Application(getOptions(this));
+  const result = app.converter.convert([this.resourcePath]);
+
+  if (result.errors && result.errors.length) {
+    result.errors.forEach(error => {
+      throw new Error(error.messageText);
+    });
+  }
+
+  const projectObject = app.serializer.projectToObject(result.project);
+  return `export default ${ JSON.stringify(projectObject) }`;
 }
